Reorganizar los ficheros y ajustar los import
